requirements.txt
flask
flask-PyMongo
pymongo
pillow